Add permission checks for all actions

+++ app/Http/Controllers/PhotoItemTagsController.php
@@ -2,15 +2,19 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StorePhotoItemTagRequest;
 use App\Models\PhotoItem;
 use App\Models\Tag;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Request;
 
 class PhotoItemTagsController extends Controller
 {
-    public function store(PhotoItem $photoItem, Request $request): JsonResponse
+    public function store(PhotoItem $photoItem, StorePhotoItemTagRequest $request): JsonResponse
     {
+        if (auth()->id() !== $photoItem->photo->user_id) {
+            abort(404);
+        }
+
         $photoItem->tags()->syncWithoutDetaching($request->input('tag_id'));
 
         return response()->json();
@@ -18,6 +22,10 @@ class PhotoItemTagsController extends Controller
 
     public function destroy(PhotoItem $photoItem, Tag $tag): JsonResponse
     {
+        if (auth()->id() !== $photoItem->photo->user_id) {
+            abort(404);
+        }
+
         $photoItem->tags()->detach($tag);
 
         return response()->json();
